# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /movie/{id}:
    x-swagger-router-controller: movie
    get:
      operationId: getOne
      description: get a movie by its ID
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralErrorResponse"
    put:
      operationId: update
      description: "update a movie by its ID"
      parameters:
        - name: id
          description: movie ID
          type: string
          in: path
          required: true
        - name: movie
          description: movie properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralSuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralErrorResponse"
    delete:
      operationId: delMovie
      description: "delete a movie by its ID"
      parameters:
        - name: id
          description: movie ID
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralSuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralErrorResponse"
...
definitions:
...
  GetMovieResponse:
    required:
      - id
      - title
      - year
    properties:
      id:
        type: string
      title: 
        type: string
      year:
        type: number
  Movie:
    type: object
    properties:
      title:
        type: string
        description: task object name
      year:
        type: number
        description: task description
    required:
      - title
      - year
  GeneralErrorResponse:
      required:
        - message
      properties:
        message:
          type: string
  GeneralSuccessResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description